+++ aaem/components/ashp_base.py
@@ -215,7 +215,7 @@ class ASHPBase (AnnualSavings):
     def calc_heat_energy_produced_per_year (self):
         """
         """
-        self.heat_energy_produced_per_year = 75*12
+        #~ self.heat_energy_produced_per_year = None
         pass # depends on child to implement
         
     def calc_heat_energy_produced_per_month (self):
@@ -299,7 +299,8 @@ class ASHPBase (AnnualSavings):
         
         factor_1 = (self.monthly_value_table['COP'] *\
                    self.monthly_value_table['% of total heating']).sum()
-        factor_2 = self.monthly_value_table['% of total heating'][self.monthly_value_table['COP'] > 0].sum()
+        factor_2 = self.monthly_value_table['% of total heating']\
+                                    [self.monthly_value_table['COP'] > 0].sum()
             
         self.average_cop = factor_1 / factor_2
       
@@ -371,7 +372,22 @@ class ASHPBase (AnnualSavings):
         """
         self.annual_heating_savings = self.baseline_heating_oil_cost - \
                                       self.proposed_ashp_operation_cost
-                                      
+                            
+    def get_fuel_total_saved (self):
+        """
+        returns the total fuel saved in gallons
+        """
+        eff = self.cd["diesel generation efficiency"]
+        proposed = self.electric_consumption/eff
+        return self.heating_oil_saved - proposed 
+                                
+    def get_total_enery_produced (self):
+        """
+        returns the total energy produced
+        """
+        return self.heat_energy_produced_per_year + \
+                self.electric_consumption * (1/constants.mmbtu_to_kWh) 
+                                     
     
     def save_component_csv (self, directory):
         """