+++ aaem/components/biomass_base.py
@@ -260,7 +260,8 @@ class BiomassBase (AnnualSavings):
         """ 
             calcualte cost of biomass fuel
         """
-        self.proposed_biomass_cost = price * self.biomass_fuel_consumed
+        self.proposed_biomass_cost = price * self.biomass_fuel_consumed + \
+                        self.maintenance_cost
         
     def calc_displaced_heating_oil_price (self):
         """
@@ -306,7 +307,21 @@ class BiomassBase (AnnualSavings):
         """
         self.annual_heating_savings = self.displaced_heating_oil_price -\
                                       self.proposed_biomass_cost 
+
+    def get_fuel_total_saved (self):
+        """
+        returns the total fuel saved in gallons
+        """
+        return self.heat_diesel_displaced
                                 
+    def get_total_enery_produced (self):
+        """
+        returns the total energy produced
+        """
+        return self.biomass_fuel_consumed / \
+                        constants.hours_per_year * \
+                        self.comp_specs['energy density']
+
     def save_component_csv (self, directory):
         """
         save the component output csv in directory