+++ aaem/components/community_buildings.py
@@ -143,6 +143,7 @@ class CommunityBuildings (AnnualSavings):
             self.diagnostics = diagnostics()
         self.cd = community_data.get_section('community')
         self.comp_specs =community_data.get_section('non-residential buildings')
+        self.percent_diesel = community_data.percent_diesel 
         self.component_name = 'non-residential buildings'
         self.forecast = forecast
         #~ print self.comp_specs['average refit cost']
@@ -277,6 +278,9 @@ class CommunityBuildings (AnnualSavings):
             self.calc_annual_net_benefit()
             
             self.calc_npv(self.cd['discount rate'], self.cd["current year"])
+            
+            ## no maintaince cost
+            self.calc_levelized_costs(0)
 
     def compare_num_buildings (self):
         """
@@ -552,9 +556,10 @@ class CommunityBuildings (AnnualSavings):
         
         
         
-        
+        # GALLONS
         self.refit_savings_fuel_Hoil_total = \
                     self.comp_specs['com building data']["Fuel Oil Post"].sum()
+        # MMBTU
         self.refit_savings_HF_total = \
             self.refit_savings_fuel_Hoil_total / constants.mmbtu_to_gal_HF
         
@@ -578,6 +583,7 @@ class CommunityBuildings (AnnualSavings):
                         self.comp_specs['com building data']["Electric"]*\
                         self.comp_specs['cohort savings multiplier']
         
+        #kWh
         self.refit_savings_kWh_total = \
                     self.comp_specs['com building data']["Electric Post"].sum()
         
@@ -599,6 +605,23 @@ class CommunityBuildings (AnnualSavings):
                 self.refit_HF_consumption*constants.mmbtu_to_gal_HF
         self.refit_kWh_consumption = self.baseline_kWh_consumption - \
                                                 self.refit_savings_kWh_total
+                                                
+    def get_fuel_total_saved (self):
+        """
+        returns the total fuel saved in gallons
+        """
+        gen_eff = self.cd["diesel generation efficiency"]
+        return self.refit_savings_fuel_Hoil_total+ \
+                self.refit_savings_kWh_total / gen_eff
+                                
+    def get_total_enery_produced (self):
+        """
+        returns the total energy saved 
+        """
+        return {'kWh': self.refit_savings_kWh_total, 
+                'MMBtu': self.refit_savings_HF_total
+               }
+        
     def calc_capital_costs (self):
         """
         pre: