+++ aaem/components/hydro.py
@@ -40,7 +40,7 @@ yaml = {'enabled': False,
 
 ## default values for yaml key/Value pairs
 yaml_defaults = {'enabled': True,
-        'lifetime': 20,
+        'lifetime': 50,
         #~ 'start year': 2017,
         }
     
@@ -255,7 +255,9 @@ def component_summary (coms, res_dir):
             diesel_red = hydro.captured_energy - hydro.lost_heat_recovery
             
             eff = hydro.cd["diesel generation efficiency"]
-                
+            
+            levelized_cost = hydro.levelized_cost_of_energy
+            break_even = hydro.break_even_cost
             #~ except AttributeError:
                 #~ offset = 0
                 #~ net_gen_hydro = 0
@@ -288,7 +290,8 @@ def component_summary (coms, res_dir):
                 
                 eff,
                 diesel_price,
-                
+                break_even,
+                levelized_cost,
                 hydro.get_NPV_benefits(),
                 hydro.get_NPV_costs(),
                 hydro.get_NPV_net_benefit(),
@@ -316,7 +319,10 @@ def component_summary (coms, res_dir):
             'Net Reduction in Heating Oil Consumption [gal]',
             'Hydro Power Reduction in Utility Diesel Consumed per year',
             'Diesel Generator Efficiency',
-            'Diesel Price - year 1 [$]',
+            'Diesel Price - year 1 [$/gal]',
+            'Break Even Diesel Price [$/gal]',
+            
+            'Levelized Cost Of Energy [$/kWh]',
             'Hydro NPV benefits [$]',
             'Hydro NPV Costs [$]',
             'Hydro NPV Net benefit [$]',
@@ -440,6 +446,12 @@ class Hydropower (AnnualSavings):
             self.calc_annual_net_benefit()
             self.calc_npv(self.cd['discount rate'], self.cd["current year"])
             
+            
+            o_m = self.net_generation_proposed * \
+                    self.cd['diesel generator o&m cost']
+        
+            self.calc_levelized_costs(o_m)
+            
     def calc_average_load (self):
         """
             calculate the average load of the system
@@ -502,6 +514,15 @@ class Hydropower (AnnualSavings):
 
     def calc_heat_recovery (self):
         """
+        caclulate heat recovery values used by component
+        pre:
+            self.percent_excess_energy is the precentage of excess energy
+        created in generation
+            self.gross_generation_proposed is a floating point value
+        post:
+            self.generation_diesel_reduction, self.lost_heat_recovery, 
+        and self.generation_diesel_reduction
+            
         """
         # %
        
@@ -540,7 +561,15 @@ class Hydropower (AnnualSavings):
 
     # Make this do stuff
     def calc_capital_costs (self):
-        """ Function Doc"""
+        """
+        calculate the captial costs
+        
+        pre:
+            the project details section of the hydro section is initlized
+        according to structure shown in load project detials
+        post:
+            self.capital_costs is the total cost of the project in dollars
+        """
         transmission_cost = \
             self.comp_specs['project details']['transmission capital cost']
         generator_cost = \
@@ -552,8 +581,16 @@ class Hydropower (AnnualSavings):
     # Make this do stuff
     def calc_annual_electric_savings (self):
         """
+        calculates the annual electric savings
+        pre:
+            self.captial_costs should be caclulated
+            self.comp_specs and self.cd should have all values nessary for 
+        this component.
+            self.diesel_prices should be intilized
+        post:
+            self.annual_electric_savings is the annual elctric savings in 
+        dollars
         """
-        
         proposed_generation_cost = self.capital_costs * \
                                 self.comp_specs['percent o&m']
                         
@@ -575,6 +612,15 @@ class Hydropower (AnnualSavings):
     # Make this do sruff. Remember the different fuel type prices if using
     def calc_annual_heating_savings (self):
         """
+        calcualte the annual heating savings
+        pre:
+            self.diesel_prices should be initilzed as a np.array of 
+        floats ($/gal)
+            self.cd should have all values initilzed as designed
+            self.captured_energy and self.lost_heat_recovery are [gallons]
+        post:
+            self.annual_heating_savings is an array or scaler in dollars
+        
         """
         price = (self.diesel_prices + self.cd['heating fuel premium'])
         
@@ -583,13 +629,19 @@ class Hydropower (AnnualSavings):
                 price
         #~ print 'self.annual_heating_savings', self.annual_heating_savings
         
-    def calc_break_even_fuel_price (self):
+    def get_fuel_total_saved (self):
         """
+        returns the total fuel saved in gallons
         """
-        #savings = (price + ofset)(energy_diff)  + price * red+ const
-        #captial_costs = price*ed + ofset*ed + price * red + const
-        #captial_costs - const - ofset*ed = price (ed+red)
-        # price = (capex - const)/ (heating_reduction + elecreduction) 
+    
+        return (self.captured_energy - self.lost_heat_recovery) + \
+                self.generation_diesel_reduction
+    
+    def get_total_enery_produced (self):
+        """
+        returns the total energy produced
+        """
+        return self.net_generation_proposed
         
     def save_component_csv (self, directory):
         """