+++ aaem/components/wind_power.py
@@ -469,7 +469,8 @@ def component_summary (coms, res_dir):
                 electric_diesel_reduction=wind.electric_diesel_reduction
                 
                 diesel_red = wind.reduction_diesel_used
-                
+                levelized_cost = wind.levelized_cost_of_energy
+                break_even = wind.break_even_cost
                 eff = wind.cd["diesel generation efficiency"]
                 
             except AttributeError:
@@ -480,6 +481,8 @@ def component_summary (coms, res_dir):
                 loss_heat = 0
                 
                 diesel_red = 0
+                levelized_cost = 0
+                break_even = 0
                 eff = wind.cd["diesel generation efficiency"]    
                 
             #~ try:
@@ -504,6 +507,8 @@ def component_summary (coms, res_dir):
                 electric_diesel_reduction,
                 eff,
                 diesel_price,
+                break_even,
+                levelized_cost,
                 wind.get_NPV_benefits(),
                 wind.get_NPV_costs(),
                 wind.get_NPV_net_benefit(),
@@ -531,7 +536,9 @@ def component_summary (coms, res_dir):
             'Heat Recovery Operational',
             'Net in Heating Oil Consumption [gal]',
             'Wind Power Reduction in Utility Diesel Consumed per year',
-            'Diesel Denerator Efficiency','Diesel Price - year 1 [$]',
+            'Diesel Denerator Efficiency','Diesel Price - year 1 [$\gal]',
+            'Break Even Diesel Price [$/gal]',
+            'Levelized Cost Of Energy [$/kWh]',
             'Wind NPV benefits [$]',
             'Wind NPV Costs [$]',
             'Wind NPV Net benefit [$]',
@@ -677,6 +684,7 @@ class WindPower(AnnualSavings):
                 self.calc_annual_net_benefit()
                 self.calc_npv(self.cd['discount rate'], self.cd["current year"])
                 #~ print self.benefit_cost_ratio
+                self.calc_levelized_costs(self.maintainance_cost)
         else:
             #~ print "wind project not feasiable"
             self.run = False
@@ -848,6 +856,18 @@ class WindPower(AnnualSavings):
         self.reduction_diesel_used = self.diesel_equiv_captured - \
                                      self.loss_heat_recovery
         #~ print 'self.reduction_diesel_used',self.reduction_diesel_used
+        
+    def get_fuel_total_saved (self):
+        """
+        returns the total fuel saved in gallons
+        """
+        return self.electric_diesel_reduction + self.reduction_diesel_used   
+    
+    def get_total_enery_produced (self):
+        """
+        returns the total energy produced
+        """
+        return self.net_generation_wind
                                      
     def calc_maintainance_cost (self):
         """